refactoring after review
