+++ src/main/java/spring/app/dao/abstraction/SongDao.java
@@ -3,8 +3,6 @@ package spring.app.dao.abstraction;
 import spring.app.model.Song;
 
 import java.sql.Timestamp;
-import java.time.LocalDate;
-import java.util.Date;
 import java.util.List;
 
 public interface SongDao extends GenericDao<Long, Song> {
@@ -15,10 +13,16 @@ public interface SongDao extends GenericDao<Long, Song> {
 
     List<Song> getAllWithGenreByGenreId(Long id);
 
-    boolean isExist(String name);
-
     List<Song> getByCreatedDateRange(Timestamp from, Timestamp to);
 
+    List<Song> getAllApproved();
+
+    List<Song> getApprovedPage(int pageNumber, int pageSize);
+
+    int getLastApprovedPageNumber(int pageSize);
+
     void bulkRemoveSongsByAuthorId(Long id);
+
+    boolean isExist(String name);
 }
