+++ src/main/java/spring/app/service/abstraction/AuthorService.java
@@ -7,24 +7,13 @@ import java.util.List;
 
 public interface AuthorService {
 
-    List<Author> getAllAuthors();
-
-    List<Author> getAllApprovedAuthors();
-
-    List<Author> getApprovedAuthorsPage(int pageNumber, int pageSize);
-
-    int getLastApprovedAuthorsPageNumber(int pageSize);
-
-    Author getByName(String name);
-
     void addAuthor(Author author);
 
-    /**
-     * Возвращает всех авторов, у которых совпадает передаваемое значение
-     *
-     * @return list<Author>
-     */
-    List<Author> findAuthorsByNameContaining(String name);
+    void updateAuthor(Author author);
+
+    void deleteAuthorById(Long id);
+
+    Author getByName(String name);
 
     /**
      * Возвращает автора по id
@@ -33,11 +22,22 @@ public interface AuthorService {
      */
     Author getById(long authorsId);
 
-    void updateAuthor(Author author);
-
-    void deleteAuthorById(Long id);
+    /**
+     * Возвращает всех авторов, у которых совпадает передаваемое значение
+     *
+     * @return list<Author>
+     */
+    List<Author> findAuthorsByNameContaining(String name);
 
     List<Author> getByCreatedDateRange(Timestamp dateFrom, Timestamp dateTo);
+
+    List<Author> getAllAuthors();
+
+    List<Author> getAllApprovedAuthors();
+
+    List<Author> getApprovedAuthorsPage(int pageNumber, int pageSize);
+
+    int getLastApprovedAuthorsPageNumber(int pageSize);
 
     boolean isExist(String name);
 }
